Extract PrimaryButton from AddressElement
