+++ paymentsheet/src/test/snapshots/images/com.stripe.android.common.ui_PrimaryButtonScreenshotTest_testDisabled[DarkTheme,CrazyAppearance,DefaultFont].png
