+++ paymentsheet/src/test/snapshots/images/com.stripe.android.common.ui_PrimaryButtonScreenshotTest_testDisabled[LightTheme,CustomAppearance,DefaultFont].png
