+++ paymentsheet/src/test/snapshots/images/com.stripe.android.common.ui_PrimaryButtonScreenshotTest_testEnabled[DarkTheme,CustomAppearance,DefaultFont].png
