+++ paymentsheet/src/test/snapshots/images/com.stripe.android.common.ui_PrimaryButtonScreenshotTest_testEnabled[LightTheme,CrazyAppearance,DefaultFont].png
